requirements.txt
gradio>=4.0.0 
openai>=1.0.0
python-dotenv